Huge refactoring to support mixed record types such as ORecordBytes with ODatabaseDocument instances.

+++ core/src/main/java/com/orientechnologies/orient/core/serialization/OBase64Utils.java
@@ -1131,7 +1131,7 @@ public class OBase64Utils {
 			else {
 				// There's a bad input character in the Base64 stream.
 				throw new OIOException(String.format("Bad Base64 input character decimal %d in array position %d",
-						((int) source[i]) & 0xFF, i));
+						(source[i]) & 0xFF, i));
 			} // end else:
 		} // each input character
 
@@ -1191,7 +1191,7 @@ public class OBase64Utils {
 		boolean dontGunzip = (options & DONT_GUNZIP) != 0;
 		if ((bytes != null) && (bytes.length >= 4) && (!dontGunzip)) {
 
-			int head = ((int) bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);
+			int head = (bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);
 			if (java.util.zip.GZIPInputStream.GZIP_MAGIC == head) {
 				java.io.ByteArrayInputStream bais = null;
 				java.util.zip.GZIPInputStream gzis = null;
